requirement.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
groq
python-dotenv